requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pymongo==4.5.0
pandas==2.1.0
prophet==1.2.4
python-dotenv==1.0.0
numpy==1.23.5
